Conflicts:
	Makefile.in
	configure.ac
	src/host.c
	src/main.c
	src/video_capture/DirectShowGrabber.cpp
	src/video_capture/deltacast.cpp
	src/video_capture/dvs.c
	src/video_capture/import.c
	src/video_compress/dxt_glsl.cpp
	src/video_compress/fastdxt.c
	src/video_compress/jpeg.cpp
	src/video_display/gl.c

+++ src/compat/inet_pton.c
@@ -96,7 +96,7 @@ u_char *dst;
                 const char *pch;
 
                 if ((pch = (char *)strchr(digits, ch)) != NULL) {
-                        uint8_t new = *tp * 10 + (pch - digits);
+                        int new = *tp * 10 + (pch - digits);
 
                         if (new > 255)
                                 return (0);
